packages/server/python-crawler/requirements.txt
PyMySQL==1.0.2
requests==2.26.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
urllib3==1.26.7
Pillow==9.2.0
kafka-python==2.0.1